mrbin.ino -- Light source now turns on when ID is successfully scanned. Also listens to serial to check if 'X' character has been sent. If the arduino reads 'X', it will turn off the light source and go back to RFID scanning mode.
CameraGUI.py -- Integrated ultrasonic sensor from arduino.py that only displays the image if there is an object in proximity. This way, we will save power and avoid heating up the Raspberry pi. If there is no object in proximity, it will send the character 'X' to the arduino to turn off the light source and go back to RFID scanning mode.

main.py -- Passed self.reader to CameraGUI.py